rename package to shutdowncheck
